requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-appconf==1.0.5
django-compressor==4.1
django-embed-video==1.4.5
django-filter==22.1
django-heroku==0.3.1
django-libsass==0.9
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
libsass==0.21.0
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
rcssmin==1.1.0
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
